+++ tests/roles/mod.rs
@@ -10,3 +10,4 @@
 mod create;
 mod read;
 mod update;
+mod delete;